validate: move the source-location helpers into location.go
validate.go carried the location resolution (the kin-error dispatch plus
the schema-field helpers) alongside the validator itself. Move all four
into location.go so one file answers how a finding gets pinned to a line,
and validate.go is just validate + the kin-error mapping. Pure move, no
behavior change.
